fix: split github merge method files under max lines
Extract the GitHub CLI runner and review-action tests so the new merge-method
files stay under the 500-line ratchet.

Amp-Thread-ID: https://ampcode.com/threads/T-01a089b0-ead1-7034-8a46-d819abd4a215
