+++ test/unit/workspace_pull_request_controller_test.dart
@@ -1,4 +1,3 @@
-import 'package:alera/src/features/pull_requests/application/forge_exception.dart';
 import 'package:alera/src/features/pull_requests/application/forge_provider.dart';
 import 'package:alera/src/features/pull_requests/application/forge_provider_registry.dart';
 import 'package:alera/src/features/pull_requests/application/pull_request_providers.dart';
@@ -397,107 +396,6 @@ void main() {
     expect(state.review?.state, HostedReviewState.closed);
     expect(state.linkedManually, isTrue);
   });
-
-  test('loads merge methods for the review base branch', () async {
-    final forge = FakeForgeProvider()
-      ..branchReview = _review(123, baseBranch: 'develop')
-      ..mergeMethods = const <ReviewMergeMethod>[ReviewMergeMethod.squash];
-    final container = _container(
-      forge: forge,
-      repo: FakeLinkedReviewRepository(),
-    );
-    addTearDown(container.dispose);
-
-    final state = await container.read(
-      workspacePullRequestControllerProvider(_scope).future,
-    );
-
-    expect(forge.mergeMethodsCalls, 1);
-    expect(forge.lastMergeMethodsBaseBranch, 'develop');
-    expect(state.mergeMethods, <ReviewMergeMethod>[ReviewMergeMethod.squash]);
-  });
-
-  test(
-    'keeps the review visible when merge methods cannot be loaded',
-    () async {
-      final forge = FakeForgeProvider()
-        ..branchReview = _review(123, baseBranch: 'main')
-        ..mergeMethodsError = const ForgeRequestFailed(
-          'GitHub did not return the repository merge methods.',
-        );
-      final container = _container(
-        forge: forge,
-        repo: FakeLinkedReviewRepository(),
-      );
-      addTearDown(container.dispose);
-
-      final state = await container.read(
-        workspacePullRequestControllerProvider(_scope).future,
-      );
-
-      expect(state.review?.number, 123);
-      expect(state.mergeMethods, isEmpty);
-      expect(state.errorMessage, isNull);
-      expect(
-        state.mergeMethodsErrorMessage,
-        'GitHub did not return the repository merge methods.',
-      );
-    },
-  );
-
-  test(
-    'refresh keeps a later review snapshot when merge methods fail',
-    () async {
-      final forge = FakeForgeProvider()
-        ..branchReview = _review(123, baseBranch: 'main')
-        ..mergeMethods = const <ReviewMergeMethod>[
-          ReviewMergeMethod.mergeCommit,
-        ]
-        ..checks = const <ReviewCheck>[
-          ReviewCheck(name: 'build', status: .completed, conclusion: .success),
-        ];
-      final container = _container(
-        forge: forge,
-        repo: FakeLinkedReviewRepository(),
-      );
-      addTearDown(container.dispose);
-
-      await container.read(
-        workspacePullRequestControllerProvider(_scope).future,
-      );
-      forge
-        ..branchReview = _review(
-          123,
-          baseBranch: 'main',
-          title: 'feat: updated',
-        )
-        ..checks = const <ReviewCheck>[
-          ReviewCheck(name: 'build', status: .completed, conclusion: .failure),
-        ]
-        ..mergeMethodsError = const ForgeRequestFailed(
-          'GitHub did not return the repository merge methods.',
-        );
-
-      await container
-          .read(workspacePullRequestControllerProvider(_scope).notifier)
-          .refresh();
-      await container.read(
-        workspacePullRequestControllerProvider(_scope).future,
-      );
-
-      final state = container
-          .read(workspacePullRequestControllerProvider(_scope))
-          .value!;
-      expect(state.review?.title, 'feat: updated');
-      expect(state.checks.single.conclusion, ReviewCheckConclusion.failure);
-      expect(state.mergeMethods, isEmpty);
-      expect(
-        state.mergeMethodsErrorMessage,
-        'GitHub did not return the repository merge methods.',
-      );
-      expect(state.errorMessage, isNull);
-    },
-  );
 
   test('blocks a merge method that the provider does not support', () async {
     final forge = FakeForgeProvider()
